+++ src/utils/tests/math_test.cpp
@@ -18,41 +18,41 @@ protected:
 // Test case for decimal_ceil function
 TEST(MathTest, DecimalCeil)
 {
-    ASSERT_EQ(decimal_ceil(3.14159f, 2), 3.15f);
-    ASSERT_EQ(decimal_ceil(1.5f, 0), 2.0f);
-    ASSERT_EQ(decimal_ceil(0.123456f, 4), 0.1235f);
+    ASSERT_EQ(decimal_ceil(3.14159, 2), 3.15);
+    ASSERT_EQ(decimal_ceil(1.5, 0), 2.0);
+    ASSERT_EQ(decimal_ceil(0.123456, 4), 0.1235);
 }
 
 // Test case for decimal_floor function
 TEST(MathTest, DecimalFloor)
 {
-    ASSERT_EQ(decimal_floor(3.14159f, 2), 3.14f);
-    ASSERT_EQ(decimal_floor(1.5f, 0), 1.0f);
-    ASSERT_EQ(decimal_floor(0.123456f, 4), 0.1234f);
+    ASSERT_EQ(decimal_floor(3.14159, 2), 3.14);
+    ASSERT_EQ(decimal_floor(1.5, 0), 1.0);
+    ASSERT_EQ(decimal_floor(0.123456, 4), 0.1234);
 }
 
 // Test case for decimal_round function
 TEST(MathTest, DecimalRound)
 {
-    ASSERT_EQ(decimal_round(3.14159f, 2), 3.14f);
-    ASSERT_EQ(decimal_round(1.5f, 0), 2.0f);
-    ASSERT_EQ(decimal_round(0.123456f, 4), 0.1235f);
+    ASSERT_EQ(decimal_round(3.14159, 2), 3.14);
+    ASSERT_EQ(decimal_round(1.5, 0), 2.0);
+    ASSERT_EQ(decimal_round(0.123456, 4), 0.1235);
 }
 
 // Test case for count_decimals function
 TEST(MathTest, CountDecimals)
 {
-    ASSERT_EQ(count_decimals(3.14159f), 5);
-    ASSERT_EQ(count_decimals(1.5f), 1);
-    ASSERT_EQ(count_decimals(0.123456f), 6);
+    ASSERT_EQ(count_decimals(3.14159), 5);
+    ASSERT_EQ(count_decimals(1.5), 1);
+    ASSERT_EQ(count_decimals(0.123456), 6);
 }
 
 // Test case for normalize function
 TEST(MathTest, Normalize)
 {
-    std::vector<float> input = {1.0f, 2.0f, 3.0f, 4.0f, 5.0f};
-    std::pair<float, float> current_range = std::make_pair(1.0f, 5.0f);
-    std::pair<float, float> new_range = std::make_pair(0.0f, 1.0f);
-    std::vector<float> expected_output = {0.0f, 0.25f, 0.5f, 0.75f, 1.0f};
+    std::vector<double> input = {1.0, 2.0, 3.0, 4.0, 5.0};
+    std::pair<double, double> current_range = std::make_pair(1.0, 5.0);
+    std::pair<double, double> new_range = std::make_pair(0.0, 1.0);
+    std::vector<double> expected_output = {0.0, 0.25, 0.5, 0.75, 1.0};
     ASSERT_EQ(normalize(input, current_range, new_range), expected_output);
 }
